Second pass at achievements
I changed the achievements system so now it relies on game events that
are sent from the server (I think all of the events propagate from
da_player.cpp so far) and are caught by the client (specifically
da_achievements.cpp).

When defining new game events you need to register them in modevents.res
- in your dab/resources directory. The only property I am defining is
"userid" as a float. I'll also upload my modevents.res file to the sync
so you can break it down.

All achievements are stat based, if it's a binary achievement I am just
setting the threshold to 1 and letting the increment trigger it. When
creating new achievements on the steamworks partner site you need to
create the stat first with the convention [achievement_name]_STAT and
there must be one stat per achievement (even for tiered grinds). This is
because the DEFINE_ACHIEVEMENT macro only takes an achievement name
(twice) as a parameter, and appends _STAT to the string to reference the
stat. Feels like the valve code is assuming on us and it's kinda shady
but it works and it's not overly messy.

Added Achievements:
Divepunch is its Own Reward - Get 1 Divepunch Kill
Rofldiver - Get 250 Divepunch Kills
Rocky Balboa - Get like a bajillion divepunch kills. (exactly 651438 - I
just mashed the numpad)
Michael Bay - Dive away from an explosion for reduced damage.
Bad Boys II - Dive away from explosions for reduced damage 250 times.
Dodge This - Kill with a headshot at point blank range

I don't want to go adding heaps of achievements until we are sure the
system works, but after that WHEEEE

I also put a bunch of comments in there for things I will come back to
later to see if I have learnt enough c++ to nut them out myself